split dassert into a separate header
