+++ src/api/wayfire/debug.hpp
@@ -6,6 +6,7 @@
 #endif
 
 #define nonull(x) ((x) ? (x) : ("nil"))
+#include <wayfire/dassert.hpp>
 #include <wayfire/util/log.hpp>
 #include <wayfire/scene.hpp>
 #include <wayfire/core.hpp>
@@ -14,42 +15,10 @@
 namespace wf
 {
 /**
- * Print the current stacktrace at runtime.
- *
- * @param fast_mode If fast_mode is true, the stacktrace will be generated
- *   using the fastest possible method. However, this means that not all
- *   information will be printed (for ex., line numbers may be missing).
- */
-void print_trace(bool fast_mode);
-
-/**
  * Dump a scenegraph to the log.
  */
 void dump_scene(scene::node_ptr root = wf::get_core().scene());
 
-/**
- * Assert that the condition is true.
- * Optionally print a message.
- * Print backtrace when the assertion fails and exit.
- */
-inline void dassert(bool condition, std::string message = "")
-{
-    if (!condition)
-    {
-        LOGE(message);
-        print_trace(false);
-        std::exit(0);
-    }
-}
-}
-
-#define DASSERT(condition) \
-    wf::dassert(condition, "Assertion failed at " __FILE__ ":" __LINE__)
-
-/* ------------------------ Logging categories -------------------------------*/
-
-namespace wf
-{
 namespace log
 {
 /**
